vul_analysis/requirements.txt
dspy-ai>=2.6.20
pandas>=2.0.0
ujson>=5.9.0
python-dotenv>=1.0.1
pydantic>=2.6.0
rich>=13.7.0
requests>=2.31.0
google-generativeai>=0.8.1
